migration to squash Headset entries that only differ by case
[#184954765]
